Split Balance component
Split Balance component to InputBalance and ToReportsButton
