+++ src/pages/Home.jsx
@@ -1,16 +1,11 @@
 import Balans from 'components/Balans/Balans';
-
 import { ToTransaction } from 'components/ToTransaction/ToTransaction';
-
-// import ExpensesIncome from 'components/ExpensesIncome/ExpensesIncome';
 
 const Home = () => {
   return (
     <div>
       <ToTransaction />
       <Balans />
-
-      {/* <ExpensesIncome expenses={expenses} income={income} /> */}
     </div>
   );
 };
